Present check for device --skip-ci

+++ config/initializers/rpush.rb
@@ -95,14 +95,18 @@ Rpush.reflect do |on|
   # You will need to replace old_id with canonical_id in your records.
   on.gcm_canonical_id do |old_id, canonical_id|
     d = ::Device.find_by_device_token(old_id)
-    d.update_attributes(device_token: canonical_id)
+    if d.present?
+      d.update_attributes(device_token: canonical_id)
+    end
   end
 
   # Called when the GCM returns a failure that indicates an invalid registration id.
   # You will need to delete the registration_id from your records.
   on.gcm_invalid_registration_id do |app, error, registration_id|
     d = ::Device.find_by_device_token(registration_id)
-    d.destroy
+    if d.present?
+      d.destroy
+    end
   end
 
   # Called when an SSL certificate will expire within 1 month.
